Always load in grsnWellSecondFloorRooms page
Since the WellWindowFake page is blank, always load in the grsnWellSecondFloorRooms page with the .age file to fill in the missing area you see through the neighborhood-instanced Gahreesens' well window (mostly in first person POV, looking up).  It's not that much extra of the age to load, and it gets rid of the ugly mistake of seeing the black background and the corners of the small sky plane through the well window.